Generate Dynamic Color Values. This function is derived from the 'ggradar' package.https://github.com/ricardo-bion/ggradar/. This function dynamically generates a vector of color values based on the number of groups. It uses RColorBrewer for smaller sets of groups and generates a gradient for larger sets.

Generate Dynamic Color Values. This function is derived from the 'ggradar' package.https://github.com/ricardo-bion/ggradar/. This function dynamically generates a vector of color values based on the number of groups. It uses RColorBrewer for smaller sets of groups and generates a gradient for larger sets.

Usage

sys_generate_color_values(num_groups)

Arguments

num_groups

The number of groups for which to generate color values.

Value

A character vector of color values.

Examples

sys_generate_color_values(5)
#> [1] "#8DD3C7" "#FFFFB3" "#BEBADA" "#FB8072" "#80B1D3"
sys_generate_color_values(20)
#>  [1] "#8DD3C7" "#B7E3BF" "#E1F3B8" "#F8F7B7" "#E0DEC5" "#C8C4D3" "#CAADC4"
#>  [8] "#E1989D" "#F78377" "#D48F90" "#A6A1B4" "#86B1CD" "#B4B2A3" "#E2B379"
#> [15] "#F1BA63" "#D6CA65" "#BAD968" "#C6D989" "#E1D3B7" "#FCCDE5"